ADD: Usar un Schema de pydantic para validar. Recomendacion de estructura y excepcion para el caso que la libreria falle
