+++ tools.py
@@ -1,0 +1,8 @@
+def valid_url(url):
+    import re
+    regex = re.compile(
+        r'^https?://'  # http:// or https://
+        r'(?:(?:[A-Z0-9](?:[A-Z0-9-]{0,61}[A-Z0-9])?\.)+[A-Z]{2,6}\.?|'  # domain...
+        r'\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3})' # ...or ip
+        r'(?:/?|[/?]\S+)$', re.IGNORECASE)
+    return url and regex.search(url) and "youtube.com" in url
